Add user authorization to task requests

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { Task } from './task.entity';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { TaskStatus } from './task-status.enum';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -15,30 +16,36 @@ export class TasksService {
 
   }
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto)
+  async getTasks(
+    filterDto: GetTasksFilterDto,
+    user: User
+  ): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user)
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const task = await this.taskRepository.findOne(id)
+  async getTaskById(
+    id: number,
+    user: User
+  ): Promise<Task> {
+    const task = await this.taskRepository.findOne({ where: { id, userId: user.id } })
 
     if (!task) throw new NotFoundException(`Task with ID ${id} not found!`);
 
     return task
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto)
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user)
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const response = await this.taskRepository.delete(id)
+  async deleteTask(id: number, user: User): Promise<void> {
+    const response = await this.taskRepository.delete({ id, userId: user.id })
 
     if (!response.affected) throw new NotFoundException(`Task with ID ${id} not found!`);
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id)
+  async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const task = await this.getTaskById(id, user)
     task.status = status
 
     await task.save()
@@ -46,38 +53,4 @@ export class TasksService {
     return task
   }
 
-
-  // getTaskById(id: string): Task {
-  //   const task = this.tasks.find(task => task.id == id);
-
-  //   if (!task) throw new NotFoundException(`Task with ID ${id} not found!`);
-
-  //   return task;
-  // }
-
-  // createTask(createTaskDto: CreateTaskDto): Task {
-  //   const { title, description } = createTaskDto;
-  //   const task: Task = {
-  //     id: uuid(),
-  //     title,
-  //     description,
-  //     status: TaskStatus.OPEN,
-  //   };
-
-  //   this.tasks.push(task);
-
-  //   return task;
-  // }
-
-  // deleteTask(id: string): void {
-  //   const found = this.getTaskById(id);
-  //   this.tasks = this.tasks.filter(task => task != found);
-  // }
-
-  // updateTaskStatus(id: string, status: TaskStatus): Task {
-  //   const task = this.getTaskById(id);
-  //   task.status = status;
-
-  //   return task;
-  // }
 }
